+++ app/src/main/java/cn/qatime/player/bean/MyExclusiveBean.java
@@ -13,7 +13,7 @@ public class MyExclusiveBean implements Serializable {
 
     /**
      * status : 1
-     * data : [{"id":829,"used_count":0,"buy_count":2,"customized_group":{"id":5,"name":"高三数学专属课","publicizes_url":{"app_info":"http://testing.qatime.cn/assets/groups/mathematics/app_info_default-af5e8db92e98e136a1a7b724b7dedd49.png","list":"http://testing.qatime.cn/assets/groups/mathematics/list_default-1b3bf1a3d82979605caea89854b60a44.png","info":"http://testing.qatime.cn/assets/groups/mathematics/info_default-3c9e978bbb618ce34ebd31defd6e2c61.png"},"subject":"数学","grade":"高三","view_tickets_count":0,"events_count":2,"closed_events_count":0,"start_at":1501257600,"end_at":1501257600}}]
+     * data : [{"id":829,"used_count":0,"buy_count":2,"customized_group":{"id":5,"name":"高三数学小班课","publicizes_url":{"app_info":"http://testing.qatime.cn/assets/groups/mathematics/app_info_default-af5e8db92e98e136a1a7b724b7dedd49.png","list":"http://testing.qatime.cn/assets/groups/mathematics/list_default-1b3bf1a3d82979605caea89854b60a44.png","info":"http://testing.qatime.cn/assets/groups/mathematics/info_default-3c9e978bbb618ce34ebd31defd6e2c61.png"},"subject":"数学","grade":"高三","view_tickets_count":0,"events_count":2,"closed_events_count":0,"start_at":1501257600,"end_at":1501257600}}]
      */
 
     private int status;
@@ -40,7 +40,7 @@ public class MyExclusiveBean implements Serializable {
          * id : 829
          * used_count : 0
          * buy_count : 2
-         * customized_group : {"id":5,"name":"高三数学专属课","publicizes_url":{"app_info":"http://testing.qatime.cn/assets/groups/mathematics/app_info_default-af5e8db92e98e136a1a7b724b7dedd49.png","list":"http://testing.qatime.cn/assets/groups/mathematics/list_default-1b3bf1a3d82979605caea89854b60a44.png","info":"http://testing.qatime.cn/assets/groups/mathematics/info_default-3c9e978bbb618ce34ebd31defd6e2c61.png"},"subject":"数学","grade":"高三","view_tickets_count":0,"events_count":2,"closed_events_count":0,"start_at":1501257600,"end_at":1501257600}
+         * customized_group : {"id":5,"name":"高三数学小班课","publicizes_url":{"app_info":"http://testing.qatime.cn/assets/groups/mathematics/app_info_default-af5e8db92e98e136a1a7b724b7dedd49.png","list":"http://testing.qatime.cn/assets/groups/mathematics/list_default-1b3bf1a3d82979605caea89854b60a44.png","info":"http://testing.qatime.cn/assets/groups/mathematics/info_default-3c9e978bbb618ce34ebd31defd6e2c61.png"},"subject":"数学","grade":"高三","view_tickets_count":0,"events_count":2,"closed_events_count":0,"start_at":1501257600,"end_at":1501257600}
          */
 
         private int id;
@@ -83,7 +83,7 @@ public class MyExclusiveBean implements Serializable {
         public static class CustomizedGroupBean {
             /**
              * id : 5
-             * name : 高三数学专属课
+             * name : 高三数学小班课
              * publicizes_url : {"app_info":"http://testing.qatime.cn/assets/groups/mathematics/app_info_default-af5e8db92e98e136a1a7b724b7dedd49.png","list":"http://testing.qatime.cn/assets/groups/mathematics/list_default-1b3bf1a3d82979605caea89854b60a44.png","info":"http://testing.qatime.cn/assets/groups/mathematics/info_default-3c9e978bbb618ce34ebd31defd6e2c61.png"}
              * subject : 数学
              * grade : 高三
@@ -102,8 +102,8 @@ public class MyExclusiveBean implements Serializable {
             private int view_tickets_count;
             private int events_count;
             private int closed_events_count;
-            private int start_at;
-            private int end_at;
+            private long start_at;
+            private long end_at;
             private String teacher_name;
 
             public int getId() {
@@ -170,19 +170,19 @@ public class MyExclusiveBean implements Serializable {
                 this.closed_events_count = closed_events_count;
             }
 
-            public int getStart_at() {
+            public long getStart_at() {
                 return start_at;
             }
 
-            public void setStart_at(int start_at) {
+            public void setStart_at(long start_at) {
                 this.start_at = start_at;
             }
 
-            public int getEnd_at() {
+            public long getEnd_at() {
                 return end_at;
             }
 
-            public void setEnd_at(int end_at) {
+            public void setEnd_at(long end_at) {
                 this.end_at = end_at;
             }
